+++ pytket/docs/changelog.rst
@@ -7,6 +7,7 @@ x.y.z (unreleased)
 Minor new features:
 * Added explicit constructors for various Python classes.
 * New ``measure_register`` method for measuring registers.
+* Added ``OpType.TK2``, a three-parameter two-qubit gate.
 * The ``GlobalisePhasedX`` transform and homonymous pass take a new optional
   ``squash`` parameter. ``squash=true`` (default) implements a new algorithm
   that significantly reduces the number of ``NPhasedX`` gates synthesised.